requirements.txt
edge_tts==6.1.8
bilibili-api-python==16.2
openai==0.28.0
langchain==0.1.11
ratelimit==2.2.1
pygame~=2.5.1
scipy==1.12
pyaudio
langchain_openai
httpx~=0.24.1
requests~=2.31.0
pydantic~=2.3.0
urllib3==1.26.18
aiofiles~=23.2.1
setuptools~=65.5.1
tiktoken~=0.5.1
python-dotenv==1.0.0
SpeechRecognition==3.10.0
browser_cookie3==0.19.1
apscheduler==3.10.4